code-catalyst/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
redis[asyncio]==5.0.1
httpx==0.25.2
typer==0.9.0
rich==13.7.0
anthropic==0.7.7
openai==1.3.9
pymongo==4.6.0
requests==2.31.0
PyGithub==2.1.1
twilio==8.10.0
